requirements.txt
# core
scipy
opencv-python
piq
adversarial-robustness-toolbox
facenet-pytorch

# plots
pandas
matplotlib
seaborn

# utils
loguru
IPython
pytest
prettytable

# api
tencentcloud-sdk-python

# misc
tensorflow